Commit the new credentials to the tenants MinIO
