+++ cluster/client.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package portal
+package cluster
 
 import (
 	"github.com/minio/minio-go/v6"
