+++ cluster/config.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package portal
+package cluster
 
 // hostConfig configuration of a host.
 type hostConfigV9 struct {
