+++ cluster/sessions.go
@@ -17,162 +17,53 @@
 package cluster
 
 import (
-	"context"
-	"database/sql"
+	"crypto/rand"
+	"crypto/sha256"
+	"encoding/base64"
 	"fmt"
-	"log"
-	"os"
-	"sync"
+	"io"
+	"time"
 
-	// postgres driver for database/sql
-	_ "github.com/lib/pq"
+	uuid "github.com/satori/go.uuid"
 )
 
-type Singleton struct {
-	Db         *sql.DB
-	tenantsCnx map[string]*sql.DB
+func CreateSession(ctx *Context, userID uuid.UUID, tenantID uuid.UUID) (*string, error) {
+	// Set query parameters
+	// Insert a new session with random string as id
+	sessionID, err := GetRandString(32, "sha256")
+	if err != nil {
+		return nil, err
+	}
+
+	query :=
+		`INSERT INTO
+				m3.provisioning.sessions ("id","user_id", "tenant_id", "occurred_at")
+			  VALUES
+				($1,$2,$3,$4)`
+	tx, err := ctx.MainTx()
+	if err != nil {
+		return nil, err
+	}
+	// Execute Query
+	_, err = tx.Exec(query, sessionID, userID, tenantID, time.Now())
+	if err != nil {
+		return nil, err
+	}
+	return &sessionID, nil
 }
 
-var instance *Singleton
-var once sync.Once
-
-// Returns a Singleton instance that keeps the connections to the Database
-func GetInstance() *Singleton {
-	once.Do(func() {
-		// Wait for the DB connection
-		ctx := context.Background()
-
-		// Get the m3 Database configuration
-		config := GetM3DbConfig()
-		db := <-ConnectToDb(ctx, config)
-
-		//build connections cache
-		cnxCache := make(map[string]*sql.DB)
-
-		instance = &Singleton{
-			Db:         db,
-			tenantsCnx: cnxCache,
-		}
-	})
-	return instance
-}
-
-// DbConfig holds the configuration to connect to a database
-type DbConfig struct {
-	// Hostname
-	Host string
-	// Port
-	Port string
-	// User
-	User string
-	// Password
-	Pwd string
-	// Database Name
-	Name string
-	// Whether SSL is enabled on the connection or not
-	Ssl bool
-	// Schema name
-	SchemaName string
-}
-
-// GetM3DbConfig returns a `DbConfig` object with the values for the database by either reading them from the environment or
-// defaulting them to a known value.
-func GetM3DbConfig() *DbConfig {
-	dbHost := "localhost"
-	if os.Getenv("DB_HOSTNAME") != "" {
-		dbHost = os.Getenv("DB_HOSTNAME")
+// GetRandString generates a random string with the defined size length
+func GetRandString(size int, method string) (string, error) {
+	rb := make([]byte, size)
+	if _, err := io.ReadFull(rand.Reader, rb); err != nil {
+		return "", err
 	}
 
-	dbPort := "5432"
-	if os.Getenv("DB_PORT") != "" {
-		dbPort = os.Getenv("DB_PORT")
+	randStr := base64.URLEncoding.EncodeToString(rb)
+	if method == "sha256" {
+		h := sha256.New()
+		h.Write([]byte(randStr))
+		randStr = fmt.Sprintf("%x", h.Sum(nil))
 	}
-
-	dbUser := "postgres"
-	if os.Getenv("DB_USER") != "" {
-		dbUser = os.Getenv("DB_USER")
-	}
-
-	dbPass := "m3meansmkube"
-	if os.Getenv("DB_PASSWORD") != "" {
-		dbPass = os.Getenv("DB_PASSWORD")
-	}
-	dbSsl := false
-	if os.Getenv("DB_SSL") != "" {
-		if os.Getenv("DB_SSL") == "true" {
-			dbSsl = true
-		}
-	}
-
-	dbName := "m3"
-	if os.Getenv("DB_NAME") != "" {
-		dbName = os.Getenv("DB_NAME")
-	}
-	return &DbConfig{
-		Host: dbHost,
-		Port: dbPort,
-		User: dbUser,
-		Pwd:  dbPass,
-		Name: dbName,
-		Ssl:  dbSsl,
-	}
-}
-
-// Creates a connection to the DB and returns it
-func ConnectToDb(ctx context.Context, config *DbConfig) chan *sql.DB {
-	ch := make(chan *sql.DB)
-	go func() {
-		defer close(ch)
-		select {
-		case <-ctx.Done():
-		default:
-			dbStr := "host=" + config.Host + " port=" + config.Port + " user=" + config.User
-			if config.Pwd != "" {
-				dbStr = dbStr + " password=" + config.Pwd
-			}
-
-			dbStr = dbStr + " dbname=" + config.Name
-			if config.Ssl {
-				dbStr = dbStr + " sslmode=enable"
-			} else {
-				dbStr = dbStr + " sslmode=disable"
-			}
-			// if a schema is sepcified, set it as the search path
-			if config.SchemaName != "" {
-				dbStr = fmt.Sprintf("%s search_path=%s", dbStr, config.SchemaName)
-			}
-
-			db, err := sql.Open("postgres", dbStr)
-			if err != nil {
-				log.Fatal(err)
-			}
-			ch <- db
-		}
-	}()
-	return ch
-}
-
-// GetTenantDB returns a database connection to the tenant being accessed, if the connection has been established
-// then it's returned from a local cache, else it's created, cached and returned.
-func (s *Singleton) GetTenantDB(tenantName string) *sql.DB {
-	// if we find the connection in the cache, return it
-	if db, ok := s.tenantsCnx[tenantName]; ok {
-		//do something here
-		return db
-	}
-	// if we reach this point, there was no connection in cache, connect and return the connection
-	ctx := context.Background()
-	// Get the tenant DB configuration
-	config := GetTenantDBConfig(tenantName)
-	tenantDbCnx := <-ConnectToDb(ctx, config)
-	s.tenantsCnx[tenantName] = tenantDbCnx
-	return s.tenantsCnx[tenantName]
-}
-
-func GetTenantDBConfig(tenantName string) *DbConfig {
-	// right now all tenants live on the same server as m3, but on a different DB
-	config := GetM3DbConfig()
-	config.Name = "tenants"
-	config.SchemaName = tenantName
-	return config
+	return randStr, nil
 }
