+++ cluster/tenants.go
@@ -17,10 +17,8 @@
 package cluster
 
 import (
-	"context"
 	"errors"
 	"fmt"
-	"log"
 	"regexp"
 
 	"github.com/golang-migrate/migrate/v4"
@@ -42,20 +40,16 @@ type AddTenantResult struct {
 }
 
 func AddTenant(name string, shortName string) error {
-	db := GetInstance().Db
-	bgCtx := context.Background()
-	// Add the tenant within a transaction in case anything goes wrong during the adding process
-	tx, err := db.BeginTx(bgCtx, nil)
+	// Start app context
+	ctx, err := NewContext(shortName)
 	if err != nil {
 		return err
 	}
 
-	ctx := NewContext(bgCtx, tx)
-
 	// register the tenant
 	tenantResult := <-InsertTenant(ctx, name, shortName)
 	if tenantResult.Error != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return tenantResult.Error
 	}
 	fmt.Println(fmt.Sprintf("Registered as tenant %s\n", tenantResult.Tenant.ID.String()))
@@ -82,37 +76,37 @@ func AddTenant(name string, shortName string) error {
 
 	if sg.Error != nil {
 		fmt.Println("There was an error adding the tenant, no storage group available.", sg.Error)
-		tx.Rollback()
+		ctx.Rollback()
 		return nil
 	}
 	// provision the tenant on that cluster
 	err = <-ProvisionTenantOnStorageGroup(ctx, tenantResult.Tenant, sg.StorageGroup)
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 	// check if we were able to provision the schema and be done running the migrations
 	err = <-tenantSchemaCh
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 	// announce the tenant on the router
 	err = <-UpdateNginxConfiguration(ctx)
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 
 	// wait for the tenant namespace to finish creating
 	err = <-namespaceCh
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 
 	// if no error happened to this point
-	err = tx.Commit()
+	err = ctx.Commit()
 	return err
 }
 
@@ -133,14 +127,21 @@ func InsertTenant(ctx *Context, tenantName string, tenantShortName string) chan 
 				m3.provisioning.tenants ("id","name","short_name")
 			  VALUES
 				($1, $2, $3)`
-		stmt, err := ctx.Prepare(query)
+		tx, err := ctx.MainTx()
 		if err != nil {
-			log.Fatal(err)
+			ch <- AddTenantResult{Error: err}
+			return
+		}
+		stmt, err := tx.Prepare(query)
+		if err != nil {
+			ch <- AddTenantResult{Error: err}
+			return
 		}
 		defer stmt.Close()
 		_, err = stmt.Exec(tenantID, tenantName, tenantShortName)
 		if err != nil {
-			log.Fatal(err)
+			ch <- AddTenantResult{Error: err}
+			return
 		}
 
 		// return result via channel
@@ -177,7 +178,11 @@ func validTenantShortName(ctx *Context, tenantShortName string) error {
 		WHERE 
 		      short_name=$1`
 	var totalCollisions int
-	row := ctx.QueryRow(checkUniqueQuery, tenantShortName)
+	tx, err := ctx.MainTx()
+	if err != nil {
+		return err
+	}
+	row := tx.QueryRow(checkUniqueQuery, tenantShortName)
 	err = row.Scan(&totalCollisions)
 	if err != nil {
 		fmt.Println(err)
@@ -279,22 +284,21 @@ func MakeBucket(tenantShortName string, bucketName string) error {
 		}
 	}
 	// Get Database connection and app Context
-	db := GetInstance().Db
-	bgCtx := context.Background()
-	tx, err := db.BeginTx(bgCtx, nil)
+	ctx, err := NewContext(tenantShortName)
 	if err != nil {
-		tx.Rollback()
 		return err
 	}
-
-	ctx := NewContext(bgCtx, tx)
 	// Get in which SG is the tenant located
 	sgt := <-GetTenantStorageGroupByShortName(ctx, tenantShortName)
+	if sgt.Error != nil {
+		ctx.Rollback()
+		return sgt.Error
+	}
 
 	// Get the credentials for a tenant
 	tenantConf, err := GetTenantConfig(sgt.Tenant.ShortName)
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 
@@ -307,7 +311,7 @@ func MakeBucket(tenantShortName string, bucketName string) error {
 		false)
 
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
 
@@ -315,10 +319,10 @@ func MakeBucket(tenantShortName string, bucketName string) error {
 	err = minioClient.MakeBucket(bucketName, "us-east-1")
 
 	if err != nil {
-		tx.Rollback()
+		ctx.Rollback()
 		return err
 	}
-	err = tx.Commit()
+	err = ctx.Commit()
 	if err != nil {
 		return nil
 	}
@@ -347,4 +351,24 @@ func createTenantNamespace(tenantShortName string) chan error {
 		}
 	}()
 	return ch
+}
+
+// getTenant gets the Tenant if it exists on the m3.provisining.tenants table
+// search is done by tenant name
+func GetTenant(tenantName string) (tenant Tenant, err error) {
+	query :=
+		`SELECT 
+				t1.id, t1.name, t1.short_name
+			FROM 
+				m3.provisioning.tenants t1
+			WHERE name=$1`
+	// non-transactional query
+	row := GetInstance().Db.QueryRow(query, tenantName)
+
+	// Save the resulted query on the User struct
+	err = row.Scan(&tenant.ID, &tenant.Name, &tenant.ShortName)
+	if err != nil {
+		return tenant, err
+	}
+	return tenant, nil
 }
