+++ cluster/utils.go
@@ -21,8 +21,12 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"path/filepath"
+	"runtime"
+
 	"strings"
 
+	"github.com/minio/minio-go/v6"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -80,4 +84,37 @@ func GetTenantConfig(shortName string) (*TenantConfiguration, error) {
 	}
 	// Build configuration
 	return &conf, nil
+}
+
+// newS3Config simply creates a new Config struct using the passed
+// parameters.
+func newS3Config(appName, url string, hostCfg *hostConfigV9) *Config {
+	// We have a valid alias and hostConfig. We populate the
+	// credentials from the match found in the config file.
+	s3Config := new(Config)
+
+	s3Config.AppName = filepath.Base(appName)
+	s3Config.AppVersion = Version
+	s3Config.AppComments = []string{filepath.Base(appName), runtime.GOOS, runtime.GOARCH}
+
+	s3Config.HostURL = url
+	if hostCfg != nil {
+		s3Config.AccessKey = hostCfg.AccessKey
+		s3Config.SecretKey = hostCfg.SecretKey
+		s3Config.Signature = hostCfg.API
+	}
+	s3Config.Lookup = toLookupType(hostCfg.Lookup)
+	return s3Config
+}
+
+// getLookupType returns the minio.BucketLookupType for lookup
+// option entered on the command line
+func toLookupType(s string) minio.BucketLookupType {
+	switch strings.ToLower(s) {
+	case "dns":
+		return minio.BucketLookupDNS
+	case "path":
+		return minio.BucketLookupPath
+	}
+	return minio.BucketLookupAuto
 }
